Moved provided and default tests into their own folders.
