+++ src/test/request/defaults/defaults.event.EventEmitter.test.ts
@@ -1,5 +1,5 @@
 // Tested Module
-import getMockReq from '../../request/request'
+import getMockReq from '../../../request/request'
 
 describe('request - Defaults from "event.EventEmitter" (accepts no arguments and return default values)', () => {
   test('req.addListener is a mocked function', () => {
