+++ src/test/request/defaults/defaults.express.test.ts
@@ -1,5 +1,5 @@
 // Tested Module
-import getMockReq from '../../request/request'
+import getMockReq from '../../../request/request'
 
 describe('request - Defaults from "express.Response" (accepts no arguments and return default values)', () => {
   test('req.params is an empty object', () => {
