+++ src/test/request/defaults/defaults.http.IncomingMessage.test.ts
@@ -1,5 +1,5 @@
 // Tested Module
-import getMockReq from '../../request/request'
+import getMockReq from '../../../request/request'
 
 describe('request - Defaults from "http.IncomingMessage" (accepts no arguments and return default values)', () => {
   test('req.aborted is a boolean', () => {
