+++ src/test/request/defaults/defaults.stream.Readable.test.ts
@@ -1,5 +1,5 @@
 // Tested Module
-import getMockReq from '../../request/request'
+import getMockReq from '../../../request/request'
 
 describe('request - Defaults from "stream.Readable" (accepts no arguments and return default values)', () => {
   test('req.readable is a boolean', () => {
