+++ src/test/response/defaults/defaults.http.OutgoingMessage.test.ts
@@ -1,5 +1,5 @@
 // Tested Module
-import getMockRes from '../../response/response'
+import getMockRes from '../../../response/response'
 
 describe('response - Defaults for "http.OutgoingMessage" (accepts no arguments and return default values)', () => {
   test('res.req is an empty object', () => {
