+++ src/test/response/defaults/defaults.http.ServerResponse.test.ts
@@ -2,7 +2,7 @@
 import { Socket } from 'net'
 
 // Tested Module
-import getMockRes from '../../response/response'
+import getMockRes from '../../../response/response'
 
 describe('response - Defaults for "http.ServerResponse" (accepts no arguments and return default values)', () => {
   test('res.statusCode is a number', () => {
