+++ src/test/response/provided/provided.stream.Writable.test.ts
@@ -1,8 +1,8 @@
 // Helpers
-import { providedFunction, providedNumber, providedBoolean } from '../helpers/provided'
+import { providedFunction, providedNumber, providedBoolean } from '../../helpers/provided'
 
 // Tested Module
-import getMockRes from '../../response/response'
+import getMockRes from '../../../response/response'
 
 describe('response - Provided for "stream.Writable" (accepts arguments and returns expected values)', () => {
   test('res.writable can be provided', () => {
